Move components into new directory structure and update imports
